requirements.txt
numpy==1.18.1
beautifulsoup4==4.8.2
opencv-contrib-python==4.1.2.30
requests==2.22.0
pillow==7.0.0